Make uninstall script not delete itself
It is not helpful if an uninstall script deletes itself, for if
anything unexpected happens during the uninstall process, one
cannot review or rerun the script.

Furthermore, if the uninstallation is done from inside a clone
of the git repository, the repository has a pending change after
the uninstall process.